+++ sdk/core/Azure.Core/src/Shared/DiagnosticScope.cs
@@ -24,6 +24,7 @@ namespace Azure.Core.Pipeline
         private readonly ActivityAdapter? _activityAdapter;
         private readonly bool _suppressNestedClientActivities;
 
+        [RequiresUnreferencedCode("The diagnosticSourceArgs are used in a call to DiagnosticSource.Write, all necessary properties need to be preserved on the type being passed in using DynamicDependency attributes.")]
 #if NETCOREAPP2_1
         internal DiagnosticScope(string scopeName, DiagnosticListener source, object? diagnosticSourceArgs, object? activitySource, ActivityKind kind, bool suppressNestedClientActivities)
 #else
@@ -136,6 +137,9 @@ namespace Azure.Core.Pipeline
         /// Marks the scope as failed.
         /// </summary>
         /// <param name="exception">The exception to associate with the failed scope.</param>
+        [UnconditionalSuppressMessage("ReflectionAnalysis", "IL2026", Justification = "The Exception being passed into this method has public properties preserved on the inner method MarkFailed." +
+            "The public property System.Exception.TargetSite.get is not compatible with trimming and produces a warning when preserving all public properties. Since we do not use this property, and" +
+            "neither does Application Insights, we can suppress the warning coming from the inner method.")]
         public void Failed(Exception? exception = default)
         {
             _activityAdapter?.MarkFailed(exception);
@@ -334,6 +338,8 @@ namespace Azure.Core.Pipeline
                 _links.Add(linkedActivity);
             }
 
+            [DynamicDependency(DynamicallyAccessedMemberTypes.PublicProperties | DynamicallyAccessedMemberTypes.PublicMethods, typeof(Activity))]
+            [DynamicDependency(DynamicallyAccessedMemberTypes.PublicProperties | DynamicallyAccessedMemberTypes.PublicMethods, typeof(DiagnosticActivity))]
             public Activity? Start()
             {
                 _currentActivity = StartActivitySourceActivity();
@@ -423,7 +429,7 @@ namespace Azure.Core.Pipeline
                     _currentActivity.Start();
                 }
 
-                _diagnosticSource.Write(_activityName + ".Start", _diagnosticSourceArgs ?? _currentActivity);
+                WriteStartEvent();
 
                 if (_displayName != null)
                 {
@@ -435,6 +441,12 @@ namespace Azure.Core.Pipeline
                 }
 
                 return _currentActivity;
+            }
+
+            [UnconditionalSuppressMessage("ReflectionAnalysis", "IL2026", Justification = "The values being passed into Write have the commonly used properties being preserved with DynamicDependency on the ActivityAdapter.Start() method, or the responsibility is on the user of this struct since the struct constructor is marked with RequiresUnreferencedCode.")]
+            private void WriteStartEvent()
+            {
+                _diagnosticSource.Write(_activityName + ".Start", _diagnosticSourceArgs ?? _currentActivity);
             }
 
             public void SetDisplayName(string displayName)
@@ -487,7 +499,8 @@ namespace Azure.Core.Pipeline
                 _currentActivity?.SetStartTime(startTime);
             }
 
-            public void MarkFailed(Exception? exception)
+            [UnconditionalSuppressMessage("ReflectionAnalysis", "IL2026", Justification = "The Exception being passed into this method has the commonly used properties being preserved with DynamicallyAccessedMemberTypes.")]
+            public void MarkFailed<[DynamicallyAccessedMembers(DynamicallyAccessedMemberTypes.PublicProperties)] T>(T exception)
             {
                 if (exception != null)
                 {
@@ -500,7 +513,7 @@ namespace Azure.Core.Pipeline
                     _currentActivity?.SetErrorStatus(exception?.ToString());
                 }
 #endif
-#if NET6_0_OR_GREATER
+#if NET6_0_OR_GREATER // SetStatus is only defined in NET 6 or greater
                 _currentActivity?.SetStatus(ActivityStatusCode.Error, exception?.ToString());
 #endif
             }
@@ -515,6 +528,7 @@ namespace Azure.Core.Pipeline
                 _tracestate = tracestate;
             }
 
+            [UnconditionalSuppressMessage("ReflectionAnalysis", "IL2026:RequiresUnreferencedCode", Justification = "The class constructor is marked with RequiresUnreferencedCode.")]
             public void Dispose()
             {
                 var activity = _currentActivity ?? _sampleOutActivity;