+++ sdk/openai/Azure.AI.OpenAI/src/Properties/AssemblyInfo.cs
@@ -3,4 +3,4 @@
 
 using Azure.Core;
 
-[assembly: Azure.Core.AzureResourceProviderNamespace("Microsoft.Insights")]
+[assembly: Azure.Core.AzureResourceProviderNamespace("Microsoft.CognitiveServices")]