+++ sdk/managednetworkfabric/Azure.ResourceManager.ManagedNetworkFabric/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 1.1.0-beta.1 (Unreleased)
+## 1.2.0-beta.1 (Unreleased)
 
 ### Features Added
 
@@ -9,6 +9,12 @@
 ### Bugs Fixed
 
 ### Other Changes
+
+## 1.1.0 (2023-09-28)
+
+### Breaking Changes
+    - New optional parameter default action added in Access Control List and Route Policy resources. 
+    - IP V4 address parameter type changed from IPAddress to String in internet gateway resource.
 
 ## 1.0.0 (2023-07-28)